Add files via upload
This is the internship given by codeclause to me.
This is a simple unit converter with converts one distance unit to another distance unit like meters to centimeters, kilometers, yards etc. as well as a temperature unit converter which converts celsius to fahrenheit , kelvin and vice versa.
For better visual presentation I used an image from online.